[ssci] Added Shipped field to READMEs
This CL adds the Shipped field (and may update the
License File field) in Chromium READMEs. Changes were
automatically created, so if you disagree with any of
them (e.g. a package is used only for testing purposes
and is not shipped), comment the suggested change and
why.

See the LSC doc at go/lsc-chrome-metadata.

Bug: b:285450740
Change-Id: I83253db9df8ecc6598ccb214139ea55775f85e24
Reviewed-on: https://chromium-review.googlesource.com/c/vulkan-deps/+/4663836
Reviewed-by: Corentin Wallez <cwallez@chromium.org>